:children_crossing: After subscribing to scheduled menu, the main menu shows up automatically.
